+++ Source/ACE.Database/Models/Shard/Biota.cs
@@ -29,13 +29,6 @@ namespace ACE.Database.Models.Shard
             BiotaPropertiesSpellBook = new HashSet<BiotaPropertiesSpellBook>();
             BiotaPropertiesString = new HashSet<BiotaPropertiesString>();
             BiotaPropertiesTextureMap = new HashSet<BiotaPropertiesTextureMap>();
-            CharacterPropertiesContract = new HashSet<CharacterPropertiesContract>();
-            CharacterPropertiesFillCompBook = new HashSet<CharacterPropertiesFillCompBook>();
-            CharacterPropertiesFriendList = new HashSet<CharacterPropertiesFriendList>();
-            CharacterPropertiesQuestRegistry = new HashSet<CharacterPropertiesQuestRegistry>();
-            CharacterPropertiesShortcutBar = new HashSet<CharacterPropertiesShortcutBar>();
-            CharacterPropertiesSpellBar = new HashSet<CharacterPropertiesSpellBar>();
-            CharacterPropertiesTitleBook = new HashSet<CharacterPropertiesTitleBook>();
         }
 
         public uint Id { get; set; }
@@ -43,7 +36,6 @@ namespace ACE.Database.Models.Shard
         public int WeenieType { get; set; }
 
         public BiotaPropertiesBook BiotaPropertiesBook { get; set; }
-        public Character Character { get; set; }
         public ICollection<BiotaPropertiesAnimPart> BiotaPropertiesAnimPart { get; set; }
         public ICollection<BiotaPropertiesAttribute> BiotaPropertiesAttribute { get; set; }
         public ICollection<BiotaPropertiesAttribute2nd> BiotaPropertiesAttribute2nd { get; set; }
@@ -66,12 +58,5 @@ namespace ACE.Database.Models.Shard
         public ICollection<BiotaPropertiesSpellBook> BiotaPropertiesSpellBook { get; set; }
         public ICollection<BiotaPropertiesString> BiotaPropertiesString { get; set; }
         public ICollection<BiotaPropertiesTextureMap> BiotaPropertiesTextureMap { get; set; }
-        public ICollection<CharacterPropertiesContract> CharacterPropertiesContract { get; set; }
-        public ICollection<CharacterPropertiesFillCompBook> CharacterPropertiesFillCompBook { get; set; }
-        public ICollection<CharacterPropertiesFriendList> CharacterPropertiesFriendList { get; set; }
-        public ICollection<CharacterPropertiesQuestRegistry> CharacterPropertiesQuestRegistry { get; set; }
-        public ICollection<CharacterPropertiesShortcutBar> CharacterPropertiesShortcutBar { get; set; }
-        public ICollection<CharacterPropertiesSpellBar> CharacterPropertiesSpellBar { get; set; }
-        public ICollection<CharacterPropertiesTitleBook> CharacterPropertiesTitleBook { get; set; }
     }
 }