+++ Source/ACE.Database/Models/Shard/CharacterPropertiesFriendList.cs
@@ -6,10 +6,10 @@ namespace ACE.Database.Models.Shard
     public partial class CharacterPropertiesFriendList
     {
         public uint Id { get; set; }
-        public uint ObjectId { get; set; }
+        public uint CharacterId { get; set; }
         public uint FriendId { get; set; }
         public uint AccountId { get; set; }
 
-        public Biota Object { get; set; }
+        public Character Character { get; set; }
     }
 }